refactor: Move sns_icon.dart into features/staff/ui dir
