+++ lib/features/staff/ui/staff_section.dart
@@ -1,6 +1,6 @@
 import 'package:confwebsite2023/core/components/responsive_widget.dart';
 import 'package:confwebsite2023/core/components/section_header.dart';
-import 'package:confwebsite2023/core/components/sns_icon.dart';
+import 'package:confwebsite2023/features/staff/ui/sns_icon.dart';
 import 'package:confwebsite2023/core/theme/app_text_style.dart';
 import 'package:confwebsite2023/core/theme/baseline_color_scheme.dart';
 import 'package:confwebsite2023/features/staff/data/staff.dart';
